Fix issue when removing/getting logs for nonexistent deployments

+++ src/routes/remove.js
@@ -18,6 +18,11 @@ module.exports = server => {
         c => c.Labels['exoframe.user'] === username && c.Names.find(n => n === `/${id}`)
       );
 
+      if (!containerInfo) {
+        reply({error: 'Container not found!'}).code(404);
+        return;
+      }
+
       const service = docker.getContainer(containerInfo.Id);
       await service.stop();
       await service.remove();
